Create task file
